+++ lib/edit_form_ui_delegate.dart
@@ -1,4 +1,5 @@
 import 'package:architecture/ui_delegate.dart';
+import 'package:flutter/material.dart';
 
 import 'edit_form_ui_state.dart';
 import 'model/edit_form_models.dart';
@@ -8,10 +9,12 @@ class EditFormUiDelegate<T>
     extends UIDelegate<EditFormUIState, EditFormUIAction> {
   final EditableModel<T> editableModel;
   final Function(T) onSubmit;
+  final bool showNonEditableFields;
 
   EditFormUiDelegate({
     required this.editableModel,
     required this.onSubmit,
+    required this.showNonEditableFields,
   });
 
   void init() {
@@ -22,6 +25,7 @@ class EditFormUiDelegate<T>
             .getFormViewTypes()
             .map((formViewType) => formViewType.getFormViewModel())
             .toList(),
+        showNonEditableFields: showNonEditableFields,
       ),
     );
   }
@@ -29,13 +33,17 @@ class EditFormUiDelegate<T>
   void submit(Map<String, String> inputValues) {
     add(const EditFormUIState.showLoading());
 
-    final map = inputValues
-        .map((key, value) => MapEntry(
-            editableModel
-                .getFormViewTypes()
-                .firstWhere((element) => element.getFormViewModel().key == key),
-            value))
-        .map((key, value) {
+    debugPrint("submit called with $inputValues");
+
+    final map = inputValues.map((key, value) {
+      debugPrint("map1 key: $key, value: $value");
+      return MapEntry(
+          editableModel
+              .getFormViewTypes()
+              .firstWhere((element) => element.getFormViewModel().key == key),
+          value);
+    }).map((key, value) {
+      debugPrint("map2 key: $key, value: $value");
       switch (key.runtimeType) {
         case FormCurrencyInputViewType:
           return MapEntry(
@@ -44,6 +52,10 @@ class EditFormUiDelegate<T>
           return MapEntry(key.getFormViewModel().key, int.parse(value));
         case FormTextInputViewType:
           return MapEntry(key.getFormViewModel().key, value);
+        case FormListTextInputViewType:
+          return MapEntry(key.getFormViewModel().key, value.split(","));
+        case FormBooleanInputViewType:
+          return MapEntry(key.getFormViewModel().key, value == "true");
         case FormSelectorInputViewType:
           return MapEntry(key.getFormViewModel().key, value);
         case FormYearPickerViewType:
@@ -56,5 +68,17 @@ class EditFormUiDelegate<T>
     });
 
     onSubmit(editableModel.fromEditJson(map));
+  }
+
+  void toggleShowNonEditableFields() {
+    add(
+      EditFormUIState.showContent(
+        viewModels: editableModel
+            .getFormViewTypes()
+            .map((formViewType) => formViewType.getFormViewModel())
+            .toList(),
+        showNonEditableFields: !showNonEditableFields,
+      ),
+    );
   }
 }