+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.2
+
+* Add preview mode to switch between preview and edit mode
+* Add Rich Text, Basic Text List and Boolean Support
+
+
+
 ## 0.0.1
 
 * Initial release.